remove file list method, add more tests
