+++ docs/changelog.txt
@@ -5,6 +5,7 @@ ver 0.5.0
 - Added "copy" mode
 - Fixed a bug that num slices won't be saved correctly for some old UE versions.
 - Refactored serialize functions
+- Removed file list method (txt will be used for batch file method)
 
 ver 0.4.5
 - Supported ShadowMapTexture2D
