+++ src/config.json
@@ -1,4 +1,4 @@
 {
-    "$desc": "UE4 version. 5.1 ~ 4.0, ff7r, and borderlans3 are available.",
-    "version": "ff7r"
+    "$desc": "UE version. 5.1 ~ 4.0, ff7r, and borderlans3 are available.",
+    "version": "4.26"
 }
