+++ src/io_util.py
@@ -66,3 +66,49 @@ def compare(file1: str, file2: str):
         i += 1
 
     raise RuntimeError(f'Not same :{i}')
+
+
+def remove_quotes(string: str) -> str:
+    if string[-1] == '\n':
+        string = string[:-1]
+    if string in ['', '"']:
+        return ''
+    if string[0] == '"':
+        string = string[1:]
+    if string[-1] == '"':
+        string = string[:-1]
+    return string
+
+
+def get_base_folder(p: str) -> tuple[str, str]:
+    folder = os.path.basename(p)
+    directory = os.path.dirname(p)
+    if folder == "":
+        folder = os.path.dirname(directory)
+        directory = os.path.basename(directory)
+    if directory == ".":
+        directory = ""
+    return directory, folder
+
+
+def get_file_list(folder: str, ext: str = None, include_base=True):
+    file_list = get_file_list_rec(folder)
+    if ext is not None:
+        file_list = [f for f in file_list if get_ext(f) in ext]
+    if include_base:
+        directory, folder = get_base_folder(folder)
+        file_list = [os.path.join(folder, file) for file in file_list]
+    else:
+        directory = folder
+    return directory, file_list
+
+
+def get_file_list_rec(folder: str) -> list[str]:
+    file_list = []
+    for file in sorted(os.listdir(folder)):
+        file_path = os.path.join(folder, file)
+        if os.path.isdir(file_path):
+            file_list += [os.path.join(file, f) for f in get_file_list_rec(file_path)]
+        else:
+            file_list.append(file)
+    return file_list
