+++ tests/test_file_path_.txt
@@ -1,0 +1,1 @@
+external\Matya-Uasset-Samples\4.27\T_Worm_A.uasset
