+++ tests/test_main.py
@@ -4,16 +4,30 @@ import shutil
 
 import pytest
 
-from . import util
-from main import main
+from . import utils_for_tests as util
+from main import main, get_config
+
+
+def base(json_args):
+    """Base function for tests."""
+    args = util.Args(json_args)
+    main(args=args, texconv=util.get_texconv())
 
 
 @pytest.mark.parametrize("json_args", util.get_test_cases("valid"))
 def test_valid_mode(json_args):
     """Test uasset i/o."""
     json_args["mode"] = "valid"
-    args = util.Args(json_args)
-    main(args=args, texconv=util.get_texconv())
+    base(json_args)
+
+
+@pytest.mark.parametrize("json_args", util.get_test_cases("valid_error"))
+def test_valid_mode_fail(json_args):
+    """Test uasset i/o."""
+    with pytest.raises(Exception) as e:
+        json_args["mode"] = "valid"
+        base(json_args)
+    assert str(e.value) == json_args["error"]
 
 
 test_file = util.get_test_cases("files")[0]
@@ -101,3 +115,28 @@ def test_offset_data(json_args):
     args.file = os.path.join("test_out", uasset)
     main(args=args, texconv=util.get_texconv())
     shutil.rmtree("test_out")
+
+
+@pytest.mark.parametrize("json_args", util.get_test_cases("save_version"))
+def test_save_version(json_args):
+    """Test save_detected_version option."""
+    args = util.Args(json_args)
+    args.mode = "check"
+    config = get_config()
+    main(args=args, config=config, texconv=util.get_texconv())
+    args = util.Args(json_args)
+    args.mode = "valid"
+    config = get_config()
+    main(args=args, config=config, texconv=util.get_texconv())
+
+
+@pytest.mark.parametrize("json_args", util.get_test_cases("dds"))
+def test_dds_io(json_args):
+    """Test dds io."""
+    base(json_args)
+
+
+@pytest.mark.parametrize("json_args", util.get_test_cases("batch"))
+def test_batch_method(json_args):
+    """Test with _file_path_.txt"""
+    base(json_args)
